Backend/app/app/requirements.txt
Django==3.0
psycopg2>=2.7,<3.0
djangorestframework
gunicorn
django-cors-headers